Update PR
package endpoints -> endpoints.framework
goal prefix endpoints -> endpoints-framework
