+++ src/main/java/com/google/cloud/tools/endpoints/framework/AbstractEndpointsWebAppMojo.java
@@ -15,7 +15,7 @@
  *
  */
 
-package com.google.cloud.tools.endpoints;
+package com.google.cloud.tools.endpoints.framework;
 
 import java.io.File;
 import java.util.List;
